+++ src/Logic/Prover.ts
@@ -12,107 +12,29 @@ import Queue from './Queue'
 import { Id, Implication } from './Types'
 import { Derivations, Proof } from './Derivations'
 
+export type { Proof } from './Derivations'
+
 // TODO: is it deduction, or derivation
 
 export type Contradiction<TheoremId = Id, PropertyId = Id> = Proof<
   TheoremId,
   PropertyId
 >
-export type Derivation<TheoremId = Id, PropertyId = Id> = {
-  property: PropertyId
-  value: boolean
-  proof: Proof<TheoremId, PropertyId>
-}
-type Result<TheoremId = Id, PropertyId = Id> =
+export type Result<TheoremId = Id, PropertyId = Id> =
   | {
-    kind: 'contradiction'
-    contradiction: Contradiction<TheoremId, PropertyId>
-  }
+      kind: 'contradiction'
+      contradiction: Contradiction<TheoremId, PropertyId>
+    }
   | { kind: 'derivations'; derivations: Derivations<TheoremId, PropertyId> }
 
-// Given a collection of implications and a collection of traits for an object,
-// find either the collection of derivable traits, or a contradiction
-export function deduceTraits<TheoremId = Id, PropertyId = Id>(
-  implications: ImplicationIndex<TheoremId, PropertyId>,
-  traits: Map<PropertyId, boolean>
-): Result<TheoremId, PropertyId> {
-  return new Prover(implications, traits).run()
-}
-
-// Given a collection of implications and a candidate formula,
-// return a proof of why the formula is unsatisfiable, if possible
-//
-// The current proof strategy is to force the formula and then look for a
-// a contradiction. Note that this does not find all possible proofs - e.g. when
-// the formula is a disjunction, `force(a | b)` does not actually force anything.
-export function disproveFormula<TheoremId = Id, PropertyId = Id>(
-  implications: ImplicationIndex<TheoremId, PropertyId>,
-  formula: Formula<PropertyId>
-): TheoremId[] | 'tautology' | undefined {
-  const prover = new Prover<TheoremId | 'given', PropertyId>(implications)
-
-  const contradiction = prover.force('given', formula)
-  if (contradiction) {
-    return formatProof(contradiction)
-  }
-
-  const result = prover.run()
-  if (result.kind === 'contradiction') {
-    return formatProof(result.contradiction)
-  }
-}
-
-// Given a collection of implications and a candidate implication,
-// attempt to derive a proof of the candidate
-export function proveTheorem<TheoremId = Id, PropertyId = Id>(
-  theorems: ImplicationIndex<TheoremId, PropertyId>,
-  when: Formula<PropertyId>,
-  then: Formula<PropertyId>
-): TheoremId[] | 'tautology' | undefined {
-  // We don't want to `disproveFormula(when + ~then)` given the current
-  // limitations of `disproveFormula` above. Instead we:
-  //
-  // * force `then`
-  // * run deductions
-  // * force `~when`
-  // * run deductions
-  //
-  // Note that this still has edges, e.g.
-  // * `A | B => C` doesn't circle back to assert `then`
-  // * `A | B => C + D` can't get started at all
-  let contradiction: Contradiction<TheoremId | 'given', PropertyId> | undefined
-  let result: Result<TheoremId | 'given', PropertyId>
-
-  const prover = new Prover<TheoremId | 'given', PropertyId>(theorems)
-  contradiction = prover.force('given', when)
-  if (contradiction) {
-    return formatProof(contradiction)
-  }
-
-  result = prover.run()
-  if (result.kind === 'contradiction') {
-    return formatProof(result.contradiction)
-  }
-
-  contradiction = prover.force('given', negate(then))
-  if (contradiction) {
-    return formatProof(contradiction)
-  }
-
-  result = prover.run()
-  if (result.kind === 'contradiction') {
-    return formatProof(result.contradiction)
-  }
-}
-
-class Prover<
+export default class Prover<
   TheoremId = Id,
   PropertyId = Id,
   Theorem extends Implication<TheoremId, PropertyId> = Implication<
     TheoremId,
     PropertyId
   >
-  > {
+> {
   private traits: Map<PropertyId, boolean>
   private derivations: Derivations<TheoremId, PropertyId>
 
@@ -143,6 +65,21 @@ class Prover<
     return { kind: 'derivations', derivations: this.derivations }
   }
 
+  force(
+    theorem: TheoremId,
+    formula: Formula<PropertyId>,
+    support: PropertyId[] = []
+  ): Contradiction<TheoremId, PropertyId> | undefined {
+    switch (formula.kind) {
+      case 'and':
+        return this.forceAnd(theorem, formula, support)
+      case 'atom':
+        return this.forceAtom(theorem, formula, support)
+      case 'or':
+        return this.forceOr(theorem, formula, support)
+    }
+  }
+
   private apply(
     implication: Theorem
   ): Contradiction<TheoremId, PropertyId> | undefined {
@@ -168,21 +105,6 @@ class Prover<
     properties: PropertyId[]
   ): Contradiction<TheoremId, PropertyId> {
     return this.derivations.expand([theorem, properties])
-  }
-
-  force(
-    theorem: TheoremId,
-    formula: Formula<PropertyId>,
-    support: PropertyId[] = []
-  ): Contradiction<TheoremId, PropertyId> | undefined {
-    switch (formula.kind) {
-      case 'and':
-        return this.forceAnd(theorem, formula, support)
-      case 'atom':
-        return this.forceAtom(theorem, formula, support)
-      case 'or':
-        return this.forceOr(theorem, formula, support)
-    }
   }
 
   private forceAtom(
@@ -227,9 +149,9 @@ class Prover<
       (
         acc:
           | {
-            falses: Formula<PropertyId>[]
-            unknown: Formula<PropertyId> | undefined
-          }
+              falses: Formula<PropertyId>[]
+              unknown: Formula<PropertyId> | undefined
+            }
           | undefined,
         sf: Formula<PropertyId>
       ) => {
@@ -265,16 +187,4 @@ class Prover<
       return this.force(theorem, result.unknown, [...support, ...falseProps])
     }
   }
-}
-
-function formatProof<TheoremId, PropertyId>(
-  proof: Proof<TheoremId | 'given', PropertyId>
-): TheoremId[] | 'tautology' {
-  const filtered: TheoremId[] = []
-  proof.theorems.forEach((id: TheoremId | 'given') => {
-    if (id !== 'given') {
-      filtered.push(id)
-    }
-  })
-  return filtered.length > 0 ? filtered : 'tautology'
 }
